requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.40
python-dotenv>=1.0.0
duckduckgo-search>=4.1.1
